requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.4
boto3==1.42.9
botocore==1.42.9
celery==5.6.2
cffi==2.0.0
charset-normalizer==3.4.4
clamd==1.0.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flower==2.0.1
humanize==4.15.0
jmespath==1.0.1
kombu==5.6.2
lxml==6.0.2
packaging==25.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
prometheus_client==0.24.0
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
pypdfium2==5.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
redis==7.1.0
ruff==0.14.2
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
tesseract==0.1.3
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
